+++ webapp/post/models/__init__.py
@@ -1,0 +1,1 @@
+from .post import Post
