+++ webapp/post/serializers/__init__.py
@@ -1,0 +1,1 @@
+from .post_serializer import PostSerializer
