+++ webapp/post/urls.py
@@ -1,5 +1,5 @@
 from rest_framework.routers import DefaultRouter
-from post.views import PostViewSet
+from post.views.post_viewset import PostViewSet
 
 router = DefaultRouter()
 router.register('', PostViewSet, basename='Post')
