+++ webapp/post/views/__init__.py
@@ -1,0 +1,1 @@
+from .post_viewset import PostViewSet
