+++ webapp/review/serializers/__init__.py
@@ -1,0 +1,1 @@
+from .review_serializer import ReviewSerializer
