+++ webapp/review/views/__init__.py
@@ -1,0 +1,2 @@
+from .review_retrieve_apiview import ReviewRetrieveAPIView
+from .review_list_apiview import ReviewListAPIView
