+++ webapp/review/views/review_retrieve_apiview.py
@@ -1,11 +1,6 @@
-from rest_framework.generics import ListAPIView, RetrieveAPIView
+from rest_framework.generics import RetrieveAPIView
 from review.models import Review
 from review.serializers import ReviewSerializer
-
-
-class ReviewListAPIView(ListAPIView):
-    queryset = Review.objects.all()
-    serializer_class = ReviewSerializer
 
 
 class ReviewRetrieveAPIView(RetrieveAPIView):
